Refactor csv creation; fix offset saving
Passing data around as dictionary is less bug-prone for later access.
Using the csv module will also help with this.
Copying the .xyz offset is not required, as offset info is written into
the .ply header.
We now only point to the sparse file for further reads if it was created
with ground and canopy, averting future problems.